fix(coworker-record): keep the corpus presentation off the client/server boundary
CI caught a real defect in my own fix, and it is worth naming precisely.

I moved presentCorpusSignals into corpus-signals.ts on the reasoning that the
module already declares "kept server-side so the client panel stays dumb". That
reasoning was right about WHERE the decision belongs and wrong about WHICH
module: corpus-signals.ts imports prisma, and ProfessionCorpusPanel is a
"use client" component. The previous import from that module was `import type`,
which erases at compile time; mine was a value import, which does not — so the
whole server graph (prisma, fs, process.cwd) got pulled into the browser bundle
and the Production Build failed with a wall of "Ecmascript file had an error"
pointing at process.cwd() in files this PR never touched.

Types cross the client boundary safely because they disappear. Functions do not.

Split into corpus-signal-presentation.ts, which imports NOTHING — verifiable
rather than asserted: the module has zero import statements, and its test suite
now runs green with no prisma client present at all, which the previous
arrangement could not do.

Verified: 13/13 across the presentation and redirect suites with no server
dependency resolvable; Production Build failure class removed at its cause.

Docs-Impact-Decision: no user-facing documentation change. The /platform route
touch is a link href pointing one hop closer to the same destination
(/platform/ai → /platform/ai/overview); users land on the identical directory
page. The corpus-panel copy change replaces a false claim with an accurate one
on an operator diagnostic tile, which the user guide does not transcribe.
