+++ apps/web/components/platform/coworker-record/ProfessionCorpusPanel.tsx
@@ -9,7 +9,10 @@
 import { DataTable, StatCard, StatusBadge, type Column } from "@/components/ui/report-kit";
 import type { Intent } from "@/components/ui/report-kit/statusColors";
 import type { CorpusGapRow, ProfessionCorpusSignals } from "@/lib/coworker-record/corpus-signals";
-import { presentCorpusSignals } from "@/lib/coworker-record/corpus-signals";
+// Value import MUST come from the dependency-free presentation module, not from
+// corpus-signals.ts — that one imports prisma, and a value import across the
+// client boundary pulls the whole server graph into the browser bundle.
+import { presentCorpusSignals } from "@/lib/coworker-record/corpus-signal-presentation";
 
 const REASON_INTENT: Record<string, Intent> = {
   unmapped: "danger",
